Move py_matter_chip_tool_adapter to chipyaml/adapters/chiptool
