--- examples/chip-tool/py_matter_chip_tool_adapter/BUILD.gn
@@ -1,38 +1,0 @@
-# Copyright (c) 2023 Project CHIP Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import("//build_overrides/build.gni")
-import("//build_overrides/chip.gni")
-
-import("//build_overrides/pigweed.gni")
-import("$dir_pw_build/python.gni")
-
-pw_python_package("matter_chip_tool_adapter") {
-  setup = [
-    "setup.py",
-    "setup.cfg",
-    "pyproject.toml",
-  ]
-
-  sources = [
-    "matter_chip_tool_adapter/__init__.py",
-    "matter_chip_tool_adapter/adapter.py",
-    "matter_chip_tool_adapter/decoder.py",
-    "matter_chip_tool_adapter/encoder.py",
-  ]
-
-  # TODO: at a future time consider enabling all (* or missing) here to get
-  #       pylint checking these files
-  static_analysis = []
-}
