--- examples/chip-tool/py_matter_chip_tool_adapter/setup.cfg
@@ -1,23 +1,0 @@
-# Copyright (c) 2023 Project CHIP Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-[metadata]
-name = matter_chip_tool_adapter
-version = 0.0.1
-author = Project CHIP Authors
-description = chip-tool adapter for matter_yamltests
-
-[options]
-packages = find:
-zip_safe = False
