+++ scripts/tests/chipyaml/adapters/chiptool/adapter.py
