+++ scripts/tests/chipyaml/adapters/chiptool/decoder.py
