+++ scripts/tests/chipyaml/adapters/chiptool/encoder.py
