+++ scripts/tests/chipyaml/relative_importer.py
@@ -29,11 +29,6 @@ except ModuleNotFoundError:
     sys.path.append(os.path.join(SCRIPT_PATH, 'tests'))
 
 try:
-    import matter_chip_tool_adapter  # noqa: F401
-except ModuleNotFoundError:
-    sys.path.append(os.path.join(EXAMPLES_PATH, 'chip-tool', 'py_matter_chip_tool_adapter'))
-
-try:
     import matter_placeholder_adapter  # noqa: F401
 except ModuleNotFoundError:
     sys.path.append(os.path.join(EXAMPLES_PATH, 'placeholder', 'py_matter_placeholder_adapter'))
