+++ scripts/tests/chipyaml/tests_tool.py
@@ -19,11 +19,12 @@ import json
 from typing import List
 
 import click
-from matter_chip_tool_adapter.decoder import MatterLog
 from runner import runner_base
 from tests_logger import TestColoredLogPrinter, WebSocketRunnerLogger
 
 from matter.yamltests.websocket_runner import WebSocketRunner, WebSocketRunnerConfig
+
+from .adapters.chiptool.decoder import MatterLog
 
 
 @click.pass_context
